修改两个都反审核 才能删除 或者 清空

+++ Src/BlueLightSPS_Web/BlueRay/src/java/com/BlueRay/mutton/service/SaleServiceImpl.java
@@ -358,7 +358,7 @@ public class SaleServiceImpl implements SaleService {
 	}
 
 	private void validatePassUnapprove(HTXX htxx) {
-		if (!"Y".equals(htxx.getSftgjhsh()) || !"Y".equals(htxx.getSftgywsh())) {
+		if (!"Y".equals(htxx.getSftgjhsh()) && !"Y".equals(htxx.getSftgywsh())) {
 			List<PCJHXX> pcjhs = planDao.getDateByHtxxId(htxx.getID());
 			for (PCJHXX pcjh : pcjhs){
 				planDao.delete(pcjh);
